Attempt auto-reconnect to Discord on presence change
If the user has Steemcord and Discord both launching on system startup,
there is no guarantee that Discord will load first, resulting in the
'Reconnect' button appearing, and presence never updating until the user
manually opens the Steemcord window and clicks 'Reconnect'.

Instead, attempt to reconnect on Presence update, with a 10 minute
cooldown to avoid unnecessary reconnect attempts if Discord is closed.